requirements.txt
numpy==2.2.1
Pillow==11.0.0
pytorch_lightning==2.1.3
PyYAML==6.0.1
PyYAML==6.0.2
torch==2.5.1
torchmetrics==1.3.0
torchvision==0.20.1
